refactor(transform): isolate shaker pattern proofs
